+++ src/Frontend/ASTBuilder.java
@@ -296,7 +296,7 @@ public class ASTBuilder extends MxBaseVisitor<ASTNode>{
     }
 
     @Override public ASTNode visitLiteral(MxParser.LiteralContext ctx) {
-        if (ctx.IntConst() != null) return new intConstNode(Integer.parseInt(ctx.IntConst().toString())
+        if (ctx.IntConst() != null) return new intConstNode(Long.parseLong(ctx.IntConst().toString())
                 , new position(ctx));
         else if (ctx.StringConst() != null) return new stringConstNode(ctx.StringConst().toString()
                 , new position(ctx));
